scripts/requirements.txt
boto3==1.40.74
botocore==1.40.74
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
idna==3.11
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
packaging==25.0
python-dateutil==2.9.0.post0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
PyYAML==6.0.2
